### 2023-03-21 #5290 @editorialbot
[PRE REVIEW]: SESAMI APP: An Accessible Interface for Surface Area Calculation of Materials from Adsorption Isotherms
**Submitting author:** @cmcp-group (<a href="http://orcid.org/0000-0002-7756-0589">Yongchul G. Chung</a>)
**Repository:** https://github.com/hjkgrp/SESAMI_web
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
