### 2023-03-21 c8c582b
joss paper minor wording changes
