+++ paper/paper.md
@@ -70,7 +70,7 @@ The monolayer loading region is assigned to a section of the isotherm where $\fr
 4. The value of $p/p_0$ calculated from BET theory, $1/(\sqrt{C}+1)$, and $p/p_0$ calculated from the third consistency rule should be equal (with ±10% tolerance).  
 5. The linear region should end at the knee of the isotherm.  
 
-Once a linear region is selected, the identified uptake value is multiplied by the molecular cross-sectional area of the adsorbate, typically derived from the bulk liquid density (16.2 Å^2^/molecule for N~2~; 14.2 Å^2^/molecule for Ar), to obtain the material's surface area, under the assumption that the adsorbate molecules only form a monolayer (Equation \ref{eq:surface_area}). The surface area obtained this way is referred as the BET area.
+Once a linear region is selected, the identified uptake value is multiplied by the molecular cross-sectional area of the adsorbate, typically derived from the bulk liquid density (16.2 Å^2^/molecule for N~2~; 14.2 Å^2^/molecule for Ar), to obtain the material's surface area, under the assumption that the adsorbate molecules only form a monolayer (Equation \ref{eq:surface_area}). The surface area obtained this way is referred to as the BET area.
 
 # Summary
 
@@ -87,7 +87,7 @@ The SESAMI web interface has extensive error handling and clearly alerts users o
 
 # Benchmarking
 
-To assess the performance of the SESAMI code in calculating surface areas from isotherms, we benchmark the SESAMI routines against other similar programs for 13 simulated and 9 experimental N~2~ isotherms obtained at 77 K for 14 metal-organic frameworks (MOFs), some of which are shown in \autoref{fig:isotherms}. Simulated isotherms are obtained from grand canonical Monte Carlo (GCMC) simulations using the open-source RASPA 2.0.47 software [@dubbeldam2016raspa], and experimental adsorption isotherms are obtained from the experimental data reported by Islamoglu and coworkers [@islamoglu2022you]. The data are then used to calculate the surface areas with the SESAMI website, BETSI [@betsi; @betsi_github], pyGAPS [@pygaps; @pygaps_docs], and BEaTmap [@beatmap]. 
+To assess the performance of the SESAMI code in calculating surface areas from isotherms, we benchmark the SESAMI routines against other similar programs for 13 simulated and 9 experimental N~2~ isotherms obtained at 77 K for 14 metal-organic frameworks (MOFs), some of which are shown in \autoref{fig:isotherms}. Simulated isotherms are obtained from grand canonical Monte Carlo (GCMC) simulations using the open-source RASPA 2.0.47 software [@dubbeldam2016raspa], and experimental adsorption isotherms are obtained from the experimental data reported by Islamoglu and coworkers [@islamoglu2022you]. The isotherms are used to calculate surface areas with the SESAMI website, BETSI [@betsi; @betsi_github], pyGAPS [@pygaps; @pygaps_docs], and BEaTmap [@beatmap]. 
 
 ![The crystal structures and isotherms of 3 of the 14 MOFs used to benchmark different isotherm to surface area codes.\label{fig:isotherms}](figures/crystal_structs_and_isotherms.tif)
 
@@ -172,6 +172,6 @@ Table: Settings used for software for isotherm to surface area calculation. All 
 
 # Acknowledgements
 
[…]
