### 2023-03-22 f5ea1d1
edits to comments in betan.py
