### 2023-03-23 #5290 @gianmarco-terrones
Hi @Kevin-Mattheus-Moerman, regarding potential reviewers, for this paper it is important that they be familiar with Python, and it would be a bonus if they have some website and chemistry/materials science background (though any technical background should be fine). Looking at the Google Spreadsheet of reviewers, some reviewers that to me seem like potential fits (in no particular order) are:

1. timokoch
2. mbagheri20
3. mhsatman
4. ulgens
5. karanprime
6. ranzhengcode
7. wexlergroup
8. ziyiyin97
9. justinvzyl
10. larsenkg
11. ziatdinovmax
12. 1mikegrn
13. elena-pascal
14. abhishekvp
15. jarrah42
16. marshallmcdonnell
17. jdalzatec
18. kinow
19. clare york
20. ma-sadeghi
21. tomtranter
22. Oldim Vilanculo
23. wjnbreu

Some of these reviewers listed as an interest "materials characterization," and the SESAMI APP is for surface area characterization so it seems like a good fit. Some of these people also mentioned web development as an interest, and the SESAMI APP is hosted on a website and has associated HTML and Javascript code. I also added the people with interest in porous media, since these are the types of materials the SESAMI code is intended for.

Also, am fairly sure ma-sadeghi contributed to a code (BEaTmap) that we benchmark in this paper, so perhaps he would be a good reviewer choice.
