### 2023-04-02 3db9c76
fixed SESAMI 1 (BET+ESW) typo in tables 2 and 4
