### 2023-05-10 #5429 @mbagheri20
SESAMI APP works well locally and online. Website design is simple and user-friendly which makes usage easy. In addition, the corresponding manuscript is well structured, and the theory part has enough details. Following are my suggestions:

**Manuscript:**
In the benchmark part, there are good tables and figures which include results for comparison with other existing approaches. But the comparison is limited to this statement “All software are in generally good agreement”. I suggest extending the comparison part a little more and explaining why the approach does not find a surface area in some cases (denoted with N/A)

**Installation:**
Since I had all the requirements pre-installed on my computer, First I checked app.py locally  (without using conda environment) and it worked well but when I used a fresh Ubuntu system to follow the installation part and used the conda environment.yml file, I was unable to install it.
There are some conflicts in yml file which should be resolved. Simple solution for me was installing without using builds and let conda decide about versions. Also, a minor thing like removing the ”prefix” in the file which is not necessary.
