### 2023-05-10 #5429 @yangmr04
> Since I had all the requirements pre-installed on my computer, First I checked app.py locally (without using conda environment) and it worked well but when I used a fresh Ubuntu system to follow the installation part and used the conda environment.yml file, I was unable to install it.

I also encountered similar situation on MacOS where I was unable to activate the environment.yml file with conda, although it was no fault of the authors but more of a flaw with conda. Same as @mbagheri20, the app ran fine after I installed all dependencies.
