### 2023-05-10 hjkgrp/SESAMI_web#27 @yangmr04
Unable to activate the conda environment locally; add a list of dependencies
https://github.com/hjkgrp/SESAMI_web/blob/3db9c768bbb895f102c2bea90c8b6948cd66a464/README.md?plain=1#L51

This is my request for changes regarding the comments made in the JOSS review https://github.com/openjournals/joss-reviews/issues/5429#issuecomment-1541914894 and https://github.com/openjournals/joss-reviews/issues/5429#issuecomment-1542257910 about unable to activate the conda environment.

First of all, as the SESAMI is designed as a web app, this is not a huge issue. The web version works beautifully!

I believe this is a known difficulty for activating conda environment on a different platform than where the code is developed. I recommend simply adding a list of all python dependencies for people who are unable activating the conda environment, or just wish to run the app without using conda. This should increase the compatibility of the SESAMI app on different operating systems, and provide a clear guide on which python packages are required to run the code.

Alternatively, one can improve the current `environment.yml`, that requires further debugging and testing. What do you think? Thanks!
