### 2023-05-12 #5429 @yangmr04
(Sorry for this long comment!)

SESAMI is an excellent tool for estimating the surface area of materials. The web-based app is elegant and lightweight, very easy to use. I have tested the code using the example provided within the SESAMI repo, as well as using [an example isotherm from the AIF repo](https://github.com/AIF-development-team/adsorptioninformationformat/blob/master/examples/NK_DUT-6_LP_N2_114PKT.aif). Everything worked as expected.

The manuscript is in general well-structured and well-written. It fulfills all JOSS publication criteria. However, I have a minor suggestion on how equations can be better introduced. Here is my proposed change, where the equations are introduced before the variables, and eliminate the use of `\ref{}`s:

```
The surface area of a material, $S$, can be calculated as
\begin{equation}\label{eq:surface_area}
S = q_m N A_m,
\end{equation}
where $q_m$ is the molar amount of adsorbate forming a monolayer per unit mass of material, $N$ is the Avogadro constant, and $A_m$ is the area taken up by a single adsorbate molecule in the monolayer.

In order to attain $q_m$, the monolayer loading region from the isotherm can be identified using the BET equation,
\begin{equation}\label{eq:BET}
\frac{p/p_0}{q(1-p/p_0)} = \frac{1}{q_m C} + \frac{C-1}{q_m C}\frac{p}{p_0},
\end{equation}
where $p$ is the vapor pressure, $p_0$ is the saturation vapor pressure, $q$ is the adsorbate loading, and $C$ is the BET constant.
```

Please also address the missing DOI picked up by the editorialbot above: https://github.com/openjournals/joss-reviews/issues/5429#issuecomment-1531488976.

@dhhagan Overall, I recommend publication on JOSS after the authors address

- the equations format as suggested above
- the missing DOI https://github.com/openjournals/joss-reviews/issues/5429#issuecomment-1531488976
- and the issue https://github.com/hjkgrp/SESAMI_web/issues/27

Thank you!
