+++ environment.txt
@@ -1,0 +1,9 @@
+python=3.8.13
+flask=1.1.2
+numpy=1.19.1
+pandas=1.1.3
+matplotlib=3.2.2
+scipy=1.5.0
+statsmodels=0.12.0
+pymongo=3.12.0
+scikit-learn=0.23.2
