+++ environment.yml
@@ -1,115 +1,14 @@
 name: SESAMI_web
 channels:
-  - plotly
   - anaconda
   - conda-forge
-  - defaults
-  - hjkgroup
 dependencies:
-  - appnope=0.1.0=py38_1001
-  - attrs=21.4.0=pyhd3eb1b0_0
-  - backcall=0.2.0=py_0
-  - black=21.12b0=pyhd8ed1ab_0
-  - blas=1.0=mkl
-  - bzip2=1.0.8=h0d85af4_4
-  - ca-certificates=2022.4.26=hecd8cb5_0
-  - click=7.1.2=py_0
-  - cycler=0.11.0=pyhd8ed1ab_0
-  - cyrus-sasl=2.1.27=h5d77f49_0
-  - dataclasses=0.8=pyhc8e2a94_3
-  - decorator=4.4.2=py_0
-  - flake8=4.0.1=pyhd8ed1ab_2
-  - flask=1.1.2=py_0
-  - flask-cors=3.0.9=py_0
-  - freetype=2.10.4=h4cff582_1
-  - importlib-metadata=4.11.3=py38h50d1736_1
-  - iniconfig=1.1.1=pyhd3eb1b0_0
-  - intel-openmp=2020.2=258
-  - ipython=7.18.1=py38h5ca1d4c_0
-  - ipython_genutils=0.2.0=py38_0
-  - itsdangerous=1.1.0=py_0
-  - jedi=0.18.1=py38h50d1736_1
-  - jinja2=2.11.2=py_0
-  - joblib=0.17.0=py_0
-  - kaleido-core=0.2.1=h0d85af4_0
-  - kiwisolver=1.3.1=py38h23ab428_0
-  - krb5=1.16.4=h1752a42_0
-  - libcxx=10.0.0=1
-  - libedit=3.1.20191231=h0678c8f_2
-  - libffi=3.4.2=h0d85af4_5
-  - libgfortran=3.0.1=h93005f0_2
-  - libntlm=1.4=h35c211d_1002
-  - libpng=1.6.37=h7cec526_2
-  - libzlib=1.2.11=h6c3fc93_1014
-  - llvm-openmp=10.0.0=h28b9765_0
-  - markupsafe=1.1.1=py38h1de35cc_1
-  - mathjax=2.7.7=0
-  - matplotlib=3.2.2=1
-  - matplotlib-base=3.2.2=py38h1300a51_1
-  - mccabe=0.6.1=py_1
-  - mkl=2019.4=233
-  - mkl-service=2.3.0=py38hfbe908c_0
-  - mkl_fft=1.2.0=py38hc64f4ea_0
-  - mkl_random=1.1.1=py38h959d312_0
-  - mypy_extensions=0.4.3=py38h50d1736_5
-  - mysql=5.7.24=h56848d4_0
-  - ncurses=6.3=h96cf925_1
-  - numpy=1.19.1=py38h3b9f5b6_0
-  - numpy-base=1.19.1=py38hcfb5961_0
-  - openssl=1.1.1o=hca72f7f_0
-  - packaging=21.3=pyhd3eb1b0_0
-  - pandas=1.1.3=py38hb1e8313_0
-  - parso=0.8.0=py_0
-  - pathspec=0.9.0=pyhd8ed1ab_0
-  - patsy=0.5.1=py38_0
-  - pexpect=4.8.0=py38_1
-  - pickleshare=0.7.5=py38_1001
-  - pip=22.0.4=pyhd8ed1ab_0
-  - platformdirs=2.5.1=pyhd8ed1ab_0
-  - plotly=5.9.0=py_0
-  - pluggy=1.0.0=py38hecd8cb5_1
-  - prompt-toolkit=3.0.8=py_0
[…]
