### 2023-05-12 581f758
fix DOI and modify equation ordering in paper
