+++ paper/paper.bib
@@ -120,7 +120,8 @@
   volume={160},
   number={07},
   pages={49--56},
-  year={2007}
+  year={2007},
+  doi={10.1016/s0167-2991(07)80008-5}
 }
 
 @article{willems2012algorithms,
