### 2023-05-12 b970e09
changed ref to autoref for equation and table references in paper
