+++ paper/paper.md
@@ -70,7 +70,7 @@ The monolayer loading region is assigned to a section of the isotherm where $\fr
 4. The value of $p/p_0$ calculated from BET theory, $1/(\sqrt{C}+1)$, and $p/p_0$ calculated from the third consistency rule should be equal (with ±10% tolerance).  
 5. The linear region should end at the knee of the isotherm.  
 
-Once a linear region is selected, the identified uptake value is multiplied by the molecular cross-sectional area of the adsorbate, typically derived from the bulk liquid density (16.2 Å^2^/molecule for N~2~; 14.2 Å^2^/molecule for Ar), to obtain the material's surface area, under the assumption that the adsorbate molecules only form a monolayer (Equation \ref{eq:surface_area}). The surface area obtained this way is referred to as the BET area.
+Once a linear region is selected, the identified uptake value is multiplied by the molecular cross-sectional area of the adsorbate, typically derived from the bulk liquid density (16.2 Å^2^/molecule for N~2~; 14.2 Å^2^/molecule for Ar), to obtain the material's surface area, under the assumption that the adsorbate molecules only form a monolayer (\autoref{eq:surface_area}). The surface area obtained this way is referred to as the BET area.
 
 # Summary
 
@@ -91,9 +91,9 @@ To assess the performance of the SESAMI code in calculating surface areas from i
 
 ![The crystal structures and isotherms of 3 of the 14 MOFs used to benchmark different isotherm to surface area codes.\label{fig:isotherms}](figures/crystal_structs_and_isotherms.tif)
 
-We find that over the set of 13 GCMC isotherms, the SESAMI machine learning model (run from the web interface) and BEaTmap have the best correlation with Zeo++ version 0.3 surface areas [@willems2012algorithms] calculated with a 1.67 Å radius probe N~2~ molecule (Tables \ref{gcmc_area_table} and \ref{corr_table}). Nevertheless, all software are in generally good agreement, underscoring the benefit of a computational approach to surface area calculation. The agreement between software is also not surprising due to the similar approach taken by most of the codes of considering multiple subsets of consecutive data points and applying checks like the Rouquerol criteria to select a linear region for BET analysis. Indeed, this agreement is also observed over the 9 experimental isotherms (Table \ref{exp_area_table}). The CIF files used to generate GCMC isotherms, benchmark isotherms, XLSX files of calculated surface areas across different software tools for both GCMC and experimental isotherms, detailed settings used for each software, and analysis scripts employed are available at <https://github.com/hjkgrp/SESAMI_web>.
[…]
