### 2023-05-12 dbe4eca
reference to software settings table
