+++ paper/paper.md
@@ -91,7 +91,7 @@ To assess the performance of the SESAMI code in calculating surface areas from i
 
 ![The crystal structures and isotherms of 3 of the 14 MOFs used to benchmark different isotherm to surface area codes.\label{fig:isotherms}](figures/crystal_structs_and_isotherms.tif)
 
-We find that over the set of 13 GCMC isotherms, the SESAMI machine learning model (run from the web interface) and BEaTmap have the best correlation with Zeo++ version 0.3 surface areas [@willems2012algorithms] calculated with a 1.67 Å radius probe N~2~ molecule (\autoref{tab:gcmc_area_table} and \autoref{tab:corr_table}). Nevertheless, all software are in generally good agreement, underscoring the benefit of a computational approach to surface area calculation. The agreement between software is also not surprising due to the similar approach taken by most of the codes of considering multiple subsets of consecutive data points and applying checks like the Rouquerol criteria to select a linear region for BET analysis. Indeed, this agreement is also observed over the 9 experimental isotherms (\autoref{tab:exp_area_table}). The CIF files used to generate GCMC isotherms, benchmark isotherms, XLSX files of calculated surface areas across different software tools for both GCMC and experimental isotherms, detailed settings used for each software, and analysis scripts employed are available at <https://github.com/hjkgrp/SESAMI_web>.
+We find that over the set of 13 GCMC isotherms, the SESAMI machine learning model (run from the web interface) and BEaTmap have the best correlation with Zeo++ version 0.3 surface areas [@willems2012algorithms] calculated with a 1.67 Å radius probe N~2~ molecule (\autoref{tab:gcmc_area_table} and \autoref{tab:corr_table}). Nevertheless, all software are in generally good agreement, underscoring the benefit of a computational approach to surface area calculation. The agreement between software is also not surprising due to the similar approach taken by most of the codes of considering multiple subsets of consecutive data points and applying checks like the Rouquerol criteria to select a linear region for BET analysis. Indeed, this agreement is also observed over the 9 experimental isotherms (\autoref{tab:exp_area_table}). The CIF files used to generate GCMC isotherms, benchmark isotherms, XLSX files of calculated surface areas across different software tools for both GCMC and experimental isotherms, detailed settings used for each software, and analysis scripts employed are available at <https://github.com/hjkgrp/SESAMI_web>. Software settings are also reported in \autoref{tab:settings_table}.
 
 : Calculated surface areas (m^2^/g) for the 13 MOFs with GCMC isotherms. Cases where a software does not find a surface area are denoted by N/A. Zeo++ calculations are conducted with the same CIF files used to generate GCMC isotherms, and a 1.67 Å radius probe N~2~ molecule, the high accuracy flag, and 2,000 Monte Carlo samples per atom are used. All other software take as input the GCMC isotherms.\label{tab:gcmc_area_table}
 
