### 2023-05-12 hjkgrp/SESAMI_web#27 @gianmarco-terrones
Hi @yangmr04 and @mbagheri20, thank you for the feedback and the kind words! With [this commit](https://github.com/hjkgrp/SESAMI_web/commit/8603bdeb776d070b7c5c742d755ebe825968c856) I believe I have addressed this issue.

I adjusted the file environment.yml as @mbagheri20 requested, removing all but the big name packages (installing these installs the other packages anyway), removing the prefix line, and removing version information in most cases to allow for easier builds. I checked, and I can make a Conda environment with this environment.yml and run the website as expected with the environment.

I also added a file environment.txt, which contains a list of python dependencies with specific versioning information for the package versions that I use in my original Conda environment when running the website locally. Furthermore, I have adjusted README.md to reflect these changes.
