### 2023-05-12 #5429 @gianmarco-terrones
Quick question, does JOSS use RGB or CMYK images?
