### 2023-05-13 hjkgrp/SESAMI_web#27 @yangmr04
Hi @gianmarco-terrones, great fix! I can confirm that now the environment.yml can be activated on my MacOS without any problem! I also get the app to run locally with native python by installing the dependent packages from environment.txt. I am happy to close this issue, but we could also wait for @mbagheri20 for further confirmation.
