### 2023-05-13 63ef68b
Extension to software comparison and explanation for when software do not find surface areas
