+++ paper/paper.md
@@ -91,7 +91,9 @@ To assess the performance of the SESAMI code in calculating surface areas from i
 
 ![The crystal structures and isotherms of 3 of the 14 MOFs used to benchmark different isotherm to surface area codes.\label{fig:isotherms}](figures/crystal_structs_and_isotherms.tif)
 
-We find that over the set of 13 GCMC isotherms, the SESAMI machine learning model (run from the web interface) and BEaTmap have the best correlation with Zeo++ version 0.3 surface areas [@willems2012algorithms] calculated with a 1.67 Å radius probe N~2~ molecule (\autoref{tab:gcmc_area_table} and \autoref{tab:corr_table}). Nevertheless, all software are in generally good agreement, underscoring the benefit of a computational approach to surface area calculation. The agreement between software is also not surprising due to the similar approach taken by most of the codes of considering multiple subsets of consecutive data points and applying checks like the Rouquerol criteria to select a linear region for BET analysis. Indeed, this agreement is also observed over the 9 experimental isotherms (\autoref{tab:exp_area_table}). The CIF files used to generate GCMC isotherms, benchmark isotherms, XLSX files of calculated surface areas across different software tools for both GCMC and experimental isotherms, detailed settings used for each software, and analysis scripts employed are available at <https://github.com/hjkgrp/SESAMI_web>. Software settings are also reported in \autoref{tab:settings_table}.
+We find that over the set of 13 GCMC isotherms, the SESAMI machine learning model (run from the web interface) and BEaTmap have the best correlation with Zeo++ version 0.3 surface areas [@willems2012algorithms] calculated with a 1.67 Å radius probe N~2~ molecule (\autoref{tab:gcmc_area_table} and \autoref{tab:corr_table}). Differences in performance across software on this benchmark set can be attributed to the differing implementations of linear region identification, or in the case of the SESAMI machine learning model, fundamental differences in how the software calculate surface areas. Nevertheless, all software are in generally good agreement, underscoring the benefit of a computational approach to surface area calculation. The agreement between software is also not surprising due to the similar approach taken by most of the codes of considering multiple subsets of consecutive data points and applying checks like the Rouquerol criteria to select a linear region for BET analysis (SESAMI 1, BETSI, BEaTmap). The agreement between software is also observed over the 9 experimental isotherms (\autoref{tab:exp_area_table}).
+
+The CIF files used to generate GCMC isotherms, benchmark isotherms, XLSX files of calculated surface areas across different software tools for both GCMC and experimental isotherms, detailed settings used for each software, and analysis scripts employed are available at <https://github.com/hjkgrp/SESAMI_web>. Software settings are also reported in \autoref{tab:settings_table}. When a software for isotherm to surface area calculation does not find a surface area, the reason can vary from no ESW minimum being found or no suitable linear region containing an ESW minimum being found in the case of SESAMI 1 (BET+ESW), to an insufficient number of data points in the chosen BET region in the case of pyGAPS.
 
 : Calculated surface areas (m^2^/g) for the 13 MOFs with GCMC isotherms. Cases where a software does not find a surface area are denoted by N/A. Zeo++ calculations are conducted with the same CIF files used to generate GCMC isotherms, and a 1.67 Å radius probe N~2~ molecule, the high accuracy flag, and 2,000 Monte Carlo samples per atom are used. All other software take as input the GCMC isotherms.\label{tab:gcmc_area_table}
 
