### 2023-05-13 #5429 @gianmarco-terrones
Hi @mbagheri20, I have adjusted the manuscript to address your comments on software comparison and N/A explanation. @cmcp-group tagging you as well since I added some sentences.

The added sentences:
"Differences in performance across software on this benchmark set can be attributed to the differing implementations of linear region identification, or in the case of the SESAMI machine learning model, fundamental differences in how the software calculate surface areas.

When a software for isotherm to surface area calculation does not find a surface area, the reason can vary from no ESW minimum being found or no suitable linear region containing an ESW minimum being found in the case of SESAMI 1 (BET+ESW), to an insufficient number of data points in the chosen BET region in the case of pyGAPS."
