+++ paper/paper.md
@@ -46,7 +46,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Accurate characterization of surface area is critical for understanding a material's properties and performance. The most widely used approach to calculate a material’s gravimetric surface area, i.e. surface area per unit mass, is the Brunauer-Emmett-Teller (BET) method [@brunauer1938adsorption]. The BET method computes the surface area of a material given the adsorption isotherm of a probe gas (i.e., N~2~ or Ar) in that material. Many researchers either obtain the BET area from commercial software that comes with measurement equipment, or perform the analyses manually on a spreadsheet, which is time-consuming and nearly impossible for some types of isotherms. Furthermore, these two approaches lead to large variability in BET-calculated areas [@betsi]. These challenges have motivated the development of programs for the automated and standardized calculation of BET areas [@sesami_1; @pygaps; @sesami_2; @betsi; @beatmap]. 
+Accurate characterization of surface area is critical for understanding a material's properties and performance. The most widely used approach to calculate a material’s gravimetric surface area, i.e. surface area per unit mass, is the Brunauer-Emmett-Teller (BET) method [@brunauer1938adsorption]. The BET method computes the surface area of a material given the adsorption isotherm of a probe gas (i.e. N~2~ or Ar) in that material. Many researchers either obtain the BET area from commercial software that comes with measurement equipment, or perform the analyses manually on a spreadsheet, which is time-consuming and nearly impossible for some types of isotherms. Furthermore, these two approaches lead to large variability in BET-calculated areas [@betsi]. These challenges have motivated the development of programs for the automated and standardized calculation of BET areas [@sesami_1; @pygaps; @sesami_2; @betsi; @beatmap]. 
 
 # BET theory background
 The surface area of a material, $S$, can be calculated as 
@@ -80,7 +80,7 @@ SESAMI 1 applies computational routines to identify suitable linear regions of a
 
 SESAMI 2 applies a machine learning (specifically, regularized linear regression with LASSO) model for the accurate surface area prediction of high surface area materials, improving on BET performance for these materials [@sesami_2]. The LASSO model uses as input the average loading in seven isotherm pressure regions as well as pairwise products of these loadings. The SESAMI 1 and 2 routines support isotherms with N~2~ and argon adsorbate at 77 K and 87 K, respectively. We note that a recent study shows that surface areas determined from N~2~ and Ar isotherms are similar, despite the 2015 IUPAC report's suggested use of Ar [@datar2022brunauer; @thommes2015physisorption]. In addition, the SESAMI 1 code supports isotherms with arbitrary user-specified adsorbates if temperature and adsorbate cross-section and saturation vapor pressure are specified.
 
[…]
