### 2023-05-19 #5429 @gianmarco-terrones
Hi @dhhagan, the version number is 1.1.0. Is the DOI generated through Zenodo? If so, I am awaiting approval from my PI to grant Zenodo access to the repo.
