### 2023-06-09 #5429 @gianmarco-terrones
@dhhagan @Kevin-Mattheus-Moerman Thank you for your efforts!

@Kevin-Mattheus-Moerman We have addressed your comments. We specified that we use an MIT License at https://zenodo.org/record/7953483 and changed the PDF such that USA is spelled out.
